Enum iroh_relay::client::ClientError

pub enum ClientError {

    Closed,
    Send,
    Receive(Error),
    ConnectTimeout,
    DialIO(Error),
    IPDisabled,
    NoLocalAddr(String),
    Hyper(Error),
    Http(Error),
    UnexpectedStatusCode(StatusCode, StatusCode),
    Upgrade(String),
    Proxy(String),
    Build(String),
    PingTimeout,
    PingAborted,
    InvalidUrl(String),
    Dns(Option<Error>),
    ActorGone,
    WebsocketError(Error),

}

Possible connection errors on the Client

Variants

Closed

The client is closed

Send

There was an error sending a packet

Receive(Error)

There was an error receiving a packet

ConnectTimeout

There was a connection timeout error

DialIO(Error)

There was an error dialing

IPDisabled

Both IPv4 and IPv6 are disabled for this relay node

NoLocalAddr(String)

No local addresses exist

Hyper(Error)

There was http server [hyper::Error]

Http(Error)

There was an http error [http::Error].

UnexpectedStatusCode(StatusCode, StatusCode)

There was an unexpected status code

Upgrade(String)

The connection failed to upgrade

Proxy(String)

The connection failed to proxy

Build(String)

The relay super::client::Client failed to build

PingTimeout

The ping request timed out

PingAborted

The ping request was aborted

InvalidUrl(String)

The given Url is invalid

Dns(Option<Error>)

There was an error with DNS resolution

ActorGone

The inner actor is gone, likely means things are shutdown.

WebsocketError(Error)

An error related to websockets, either errors with parsing ws messages or the handshake