iroh_relay::dns
Module node_info

Available on non-wasm_browser only.

Support for handling DNS resource records for dialing by NodeId.

Dialing by NodeId is supported by iroh nodes publishing Pkarr records to DNS servers or the Mainline DHT. This module supports creating and parsing these records.

DNS records are published under the following names:

_iroh.<z32-node-id>.<origin-domain> TXT

_iroh is the record name as defined by IROH_TXT_NAME.

<z32-node-id> is the z-base-32 encoding of the NodeId.

<origin-domain> is the domain name of the publishing DNS server, N0_DNS_NODE_ORIGIN_PROD is the server operated by number0 for production. N0_DNS_NODE_ORIGIN_STAGING is the server operated by number0 for testing.

TXT is the DNS record type.

The returned TXT records must contain a string value of the form key=value as defined in RFC1464. The following attributes are defined:

relay=<url>: The home RelayUrl of this node.

addr=<addr> <addr>: A space-separated list of sockets addresses for this iroh node. Each address is an IPv4 or IPv6 address with a port.

Structs

NodeInfo

Information about the iroh node contained in an IROH_TXT_NAME TXT resource record.

Constants

IROH_TXT_NAME

The DNS name for the iroh TXT record.

Functions

from_z32

Parses a NodeId from [z-base-32] encoding.

node_id_from_domain_name

Parses a NodeId from a DNS domain name.

to_z32

Encodes a NodeId in [z-base-32] encoding.