iroh_relay

Module protos

Protocols used by the iroh-relay

Modules

common

Common types between the super::handshake and super::relay protocols.

disco

This module exports looks_like_disco_wrapper as the only disco-related relay functionality.

handshake

Implements the handshake protocol that authenticates and authorizes clients connecting to the relays.

relay

This module implements the send/recv relaying protocol.

streams

Implements logic for abstracting over a websocket stream that allows sending only [Bytes]-based messages.