iroh_relay::protos

Module send_recv

This module implements the relaying protocol used by the server and client.

Protocol flow:

Login:

client connects

-> client sends FrameType::ClientInfo

Steady state:

server occasionally sends FrameType::KeepAlive (or FrameType::Ping)

client responds to any FrameType::Ping with a FrameType::Pong

clients sends FrameType::SendPacket

server then sends FrameType::RecvPacket to recipient

Structs

Datagrams

One or multiple datagrams being transferred via the relay.

Enums

ClientToServerMsg

TODO(matheus23): Docs

RecvError

Protocol send errors.

SendError

Protocol send errors.

ServerToClientMsg

TODO(matheus23): Docs The messages received from a framed relay stream.

Constants

MAX_PACKET_SIZE

The maximum size of a packet sent over relay. (This only includes the data bytes visible to magicsock, not including its on-wire framing overhead)

MAX_PAYLOAD_SIZE

Maximum size a datagram payload is allowed to be.