iroh_relay::protos::streams

Type Alias StreamError

pub type StreamError = Error;

Available on non-wasm_browser only.

Error type for WebSocket stream operations.

This type alias represents errors that can occur during WebSocket communication. The underlying error type depends on the platform:

On non-browser platforms: tokio_websockets::Error

On browser WASM: ws_stream_wasm::WsErr

Aliased Type

pub enum StreamError {
    AlreadyClosed,
    CannotResolveHost,
    NoUriConfigured,
    DisallowedHeader,
    Protocol(ProtocolError),
    PayloadTooLong {
        len: usize,
        max_len: usize,
    },
    Io(Error),
    InvalidDNSName(InvalidDnsNameError),
    Rustls(Error),
    UnsupportedScheme,
    Upgrade(Error),
}

Variants

AlreadyClosed

Attempted to read from or write to a closed stream.

CannotResolveHost

DNS lookup failed.

NoUriConfigured

Attempted to connect a client to a remote without configured URI.

DisallowedHeader

Attempted to add a disallowed header.

Protocol(ProtocolError)

WebSocket protocol violation.

PayloadTooLong

Payload length limit was exceeded.

Fields

len: usize

max_len: usize

Io(Error)

I/O error.

InvalidDNSName(InvalidDnsNameError)

Attempted to connect to an invalid DNS name.

Rustls(Error)

A general rustls error.

UnsupportedScheme

An unsupported, i.e. not ws or wss, or no URI scheme was specified.

Upgrade(Error)

The HTTP/1.1 Upgrade failed.