Module iroh_relay::protos::stun

STUN packets sending and receiving.

Modules

methods

STUN Methods Registry

Structs

MessageDecoder

Class used to decode STUN messages

StunDecodeError

Describes an error decoding a StunMessage

TransactionId

The transaction ID is a 96-bit identifier, used to uniquely identify STUN transactions. It primarily serves to correlate requests with responses, though it also plays a small role in helping to prevent certain types of attacks. The server also uses the transaction ID as a key to identify each transaction uniquely across all clients.

Enums

Error

Errors that can occur when handling a STUN packet.

MessageClass

The STUN message class. Although there are four message classes, there are only two types of transactions in STUN: request/response transactions (which consist of a request message and a response message) and indication transactions (which consist of a single indication message). Response classes are split into error and success responses to aid in quickly processing the STUN message.

StunAttribute

STUN Attributes that can be attached to a StunMessage

Functions

is

Reports whether b is a STUN message.

parse_binding_request

Parses a STUN binding request.

parse_response

Parses a successful binding response STUN packet. The IP address is extracted from the XOR-MAPPED-ADDRESS attribute.

request

Generates a binding request STUN packet.

response

Generates a binding response.