iroh_relay::server

Module client

Available on crate feature server only.

The server-side representation of an ongoing client relaying connection.

Structs

Client

The Server side representation of a Client’s connection.

Config

Configuration for a Client.

ForwardPacketError

Error returned when forwarding a packet to a client fails.

Enums

HandleFrameError

Error when handling an incoming frame from a client.

RunError

Run error

WriteFrameError

Error when writing a frame to a client.