iroh_relay::server
Struct Metrics

pub struct Metrics {

    pub bytes_sent: Counter,
    pub bytes_recv: Counter,
    pub send_packets_sent: Counter,
    pub send_packets_recv: Counter,
    pub send_packets_dropped: Counter,
    pub disco_packets_sent: Counter,
    pub disco_packets_recv: Counter,
    pub disco_packets_dropped: Counter,
    pub other_packets_sent: Counter,
    pub other_packets_recv: Counter,
    pub other_packets_dropped: Counter,
    pub got_ping: Counter,
    pub sent_pong: Counter,
    pub unknown_frames: Counter,
    pub frames_rx_ratelimited_total: Counter,
    pub conns_rx_ratelimited_total: Counter,
    pub accepts: Counter,
    pub disconnects: Counter,
    pub unique_client_keys: Counter,
    pub websocket_accepts: Counter,
    pub relay_accepts: Counter,

}

Available on crate feature server only.

Metrics tracked for the relay server

Fields
bytes_sent: Counter

Bytes sent from a FrameType::SendPacket

bytes_recv: Counter

Bytes received from a FrameType::SendPacket

send_packets_sent: Counter

FrameType::SendPacket sent, that are not disco messages

send_packets_recv: Counter

FrameType::SendPacket received, that are not disco messages

send_packets_dropped: Counter

FrameType::SendPacket dropped, that are not disco messages

disco_packets_sent: Counter

FrameType::SendPacket sent that are disco messages

disco_packets_recv: Counter

FrameType::SendPacket received that are disco messages

disco_packets_dropped: Counter

FrameType::SendPacket dropped that are disco messages

other_packets_sent: Counter

Packets of other FrameTypes sent

other_packets_recv: Counter

Packets of other FrameTypes received

other_packets_dropped: Counter

Packets of other FrameTypes dropped

got_ping: Counter

Number of FrameType::Pings received

sent_pong: Counter

Number of FrameType::Pongs sent

unknown_frames: Counter

Number of FrameType::Unknown received

frames_rx_ratelimited_total: Counter

Number of frames received from client connection which have been rate-limited.

conns_rx_ratelimited_total: Counter

Number of client connections which have had any frames rate-limited.

accepts: Counter

Number of connections we have accepted

disconnects: Counter

Number of connections we have removed because of an error

unique_client_keys: Counter

Number of unique client keys per day

websocket_accepts: Counter

Number of accepted websocket connections

relay_accepts: Counter

Number of accepted โiroh derp httpโ connection upgrades