iroh_relay
Module time

Sleep and timeout utilities that work natively (via tokio) and in the browser.

Structs

DurationNon-wasm_browser

A Duration type to represent a span of time, typically used for system timeouts.

ElapsedNon-wasm_browser

Errors returned by Timeout.

InstantNon-wasm_browser

A measurement of a monotonically nondecreasing clock. Opaque and useful only with Duration.

SleepNon-wasm_browser

Future returned by sleep and sleep_until.

TimeoutNon-wasm_browser

Future returned by timeout and timeout_at.

Functions

sleepNon-wasm_browser

Waits until duration has elapsed.

sleep_untilNon-wasm_browser

Waits until deadline is reached.

timeoutNon-wasm_browser

Requires a Future to complete before the specified duration has elapsed.