Module iroh_relay::protos::relay

This module implements the relaying protocol used by the server and client.

Protocol flow:

Login:

client connects

-> client sends FrameType::ClientInfo

Steady state:

server occasionally sends FrameType::KeepAlive (or FrameType::Ping)

client responds to any FrameType::Ping with a FrameType::Pong

clients sends FrameType::SendPacket

server then sends FrameType::RecvPacket to recipient

Constants

MAX_PACKET_SIZE

The maximum size of a packet sent over relay. (This only includes the data bytes visible to magicsock, not including its on-wire framing overhead)